datasets
It a collection of data related to a particular topic, theme, or industry. Datasets include different types of information, such as numbers, text, images, videos, and audio, and can be stored in various formats, such as CSV, JSON, or SQL. So, a dataset typically involves structured data for a specific purpose and is related to the same subject.

You can use datasets to conduct market research, analyze competitors, compare prices, identify and study trends, or train machine learning models. These are just a few examples, and datasets are useful in various areas and situations.